i must have forgotten to commit this --- don't redownload LFS files on update

+++ dojo_plugin/utils/dojo.py
@@ -186,6 +186,8 @@ def dojo_initialize_files(data, dojo_dir):
         abs_path = dojo_dir / rel_path
         assert not abs_path.is_symlink(), f"{rel_path} is a symbolic link!"
         if dojo_file["type"] == "download":
+            if abs_path.exists():
+                continue
             abs_path.parent.mkdir(parents=True, exist_ok=True)
             urllib.request.urlretrieve(dojo_file["url"], str(abs_path))
             assert abs_path.stat().st_size >= 50*1024*1024, f"{rel_path} is small enough to fit into git ({abs_path.stat().st_size} bytes) --- put it in the repository!"
